refactor(bitcoin-client): separate bitcoin-client module from rooch-common
- Moved `bitcoin_client` to a new `crates/bitcoin-client` directory
- Updated references in relevant modules accordingly
- Adjusted dependencies in `Cargo.toml` and `Cargo.lock`
