+++ crates/bitcoin-client/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "bitcoin-client"
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+publish.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+async-trait = { workspace = true }
+bitcoin = { workspace = true }
+bitcoincore-rpc = { workspace = true }
+coerce = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tokio = { features = ["full"], workspace = true }
+tracing = { workspace = true }
