+++ crates/bitcoin-client/src/actor/mod.rs
@@ -1,6 +1,5 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod actor;
+pub mod client;
 pub mod messages;
-pub mod proxy;
