+++ crates/bitcoin-client/src/lib.rs
@@ -1,5 +1,5 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod fs;
-pub mod utils;
+pub mod actor;
+pub mod proxy;
