+++ crates/rooch-common/Cargo.toml
@@ -12,13 +12,4 @@ repository = { workspace = true }
 rust-version = { workspace = true }
 
 [dependencies]
-anyhow = { workspace = true }
-async-trait = { workspace = true }
-bitcoin = { workspace = true }
-bitcoincore-rpc = { workspace = true }
-coerce = { workspace = true }
 libc = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tokio = { features = ["full"], workspace = true }
-tracing = { workspace = true }
