+++ crates/rooch-pipeline-processor/src/actor/processor.rs
@@ -8,13 +8,13 @@ use crate::metrics::PipelineProcessorMetrics;
 use anyhow::{Error, Result};
 use async_trait::async_trait;
 use bitcoin::hashes::Hash;
+use bitcoin_client::proxy::BitcoinClientProxy;
 use coerce::actor::{context::ActorContext, message::Handler, Actor, LocalActorRef};
 use function_name::named;
 use moveos::moveos::VMPanicError;
 use moveos_types::state::StateChangeSetExt;
 use moveos_types::transaction::VerifiedMoveOSTransaction;
 use prometheus::Registry;
-use rooch_common::bitcoin_client::proxy::BitcoinClientProxy;
 use rooch_da::actor::messages::AppendTransactionMessage;
 use rooch_da::proxy::DAServerProxy;
 use rooch_db::RoochDB;
