+++ crates/rooch-relayer/Cargo.toml
@@ -30,7 +30,7 @@ moveos-eventbus = { workspace = true }
 metrics = { workspace = true }
 
 rooch-types = { workspace = true }
-rooch-common = { workspace = true }
+bitcoin-client = { workspace = true }
 rooch-config = { workspace = true }
 rooch-executor = { workspace = true }
 rooch-pipeline-processor = { workspace = true }
