+++ crates/rooch-relayer/src/actor/relayer.rs
@@ -6,13 +6,13 @@ use super::ethereum_relayer::EthereumRelayer;
 use crate::actor::relayer_proxy::RelayerProxy;
 use anyhow::Result;
 use async_trait::async_trait;
+use bitcoin_client::actor::client::BitcoinClientActor;
+use bitcoin_client::actor::messages::RelayTick;
+use bitcoin_client::proxy::BitcoinClientProxy;
 use coerce::actor::{context::ActorContext, message::Handler, Actor, LocalActorRef};
 use move_core_types::vm_status::KeptVMStatus;
 use moveos_eventbus::bus::EventData;
 use moveos_types::module_binding::MoveFunctionCaller;
-use rooch_common::bitcoin_client::actor::BitcoinClientActor;
-use rooch_common::bitcoin_client::messages::RelayTick;
-use rooch_common::bitcoin_client::proxy::BitcoinClientProxy;
 use rooch_config::{BitcoinRelayerConfig, EthereumRelayerConfig};
 use rooch_event::actor::{EventActor, EventActorSubscribeMessage};
 use rooch_event::event::ServiceStatusEvent;
