+++ crates/rooch-rpc-server/Cargo.toml
@@ -39,7 +39,7 @@ moveos-eventbus = { workspace = true }
 raw-store = { workspace = true }
 metrics = { workspace = true }
 
-rooch-common = { workspace = true }
+bitcoin-client = { workspace = true }
 rooch-config = { workspace = true }
 rooch-types = { workspace = true }
 rooch-executor = { workspace = true }
