+++ crates/rooch-rpc-server/src/lib.rs
@@ -11,14 +11,14 @@ use crate::service::metrics::ServiceMetrics;
 use crate::service::rpc_service::RpcService;
 use anyhow::{ensure, Error, Result};
 use axum::http::{HeaderValue, Method};
+use bitcoin_client::actor::client::BitcoinClientConfig;
+use bitcoin_client::actor::messages::RelayTick;
+use bitcoin_client::proxy::BitcoinClientProxy;
 use coerce::actor::scheduler::timer::Timer;
 use coerce::actor::{system::ActorSystem, IntoActor};
 use jsonrpsee::RpcModule;
 use moveos_eventbus::bus::EventBus;
 use raw_store::errors::RawStoreError;
-use rooch_common::bitcoin_client::actor::BitcoinClientConfig;
-use rooch_common::bitcoin_client::messages::RelayTick;
-use rooch_common::bitcoin_client::proxy::BitcoinClientProxy;
 use rooch_config::da_config::derive_genesis_namespace;
 use rooch_config::server_config::ServerConfig;
 use rooch_config::settings::PROPOSER_CHECK_INTERVAL;
